Add reste cotisations to ba
