+++ openfisca_nouvelle_caledonie/variables/prelevements_obligatoires/impot_revenu/revenus_imposbales/non_salarie/cotisations.py
@@ -2,6 +2,9 @@
 
 from openfisca_core.model_api import *
 from openfisca_nouvelle_caledonie.entities import Person as Individu
+from openfisca_nouvelle_caledonie.variables.prelevements_obligatoires.impot_revenu.revenus_imposbales.non_salarie import (
+    get_multiple_and_plafond_cafat_cotisation
+    )
 
 # • Indiquez lignes QA, QB, QC vos cotisations de retraite (en tant que chef d’entre-
 # prise) dans la limite du plafond, soit 3 776 500 F.
@@ -49,10 +52,14 @@ class cotisations_non_salarie(Variable):
 
     def formula(individu, period, parameters):
         multiple, plafond_cafat = get_multiple_and_plafond_cafat_cotisation(period, parameters)
-        return (
-            min_(individu("cotisations_retraite_exploitant", period), multiple * plafond_cafat)
-            + individu("cotisations_ruamm_mutuelle_ccs_exploitant", period)
+        cotisations_non_salarie = max_(
+            (
+                min_(individu("cotisations_retraite_exploitant", period), multiple * plafond_cafat)
+                + individu("cotisations_ruamm_mutuelle_ccs_exploitant", period)
+                ),
+            0
             )
+        return cotisations_non_salarie
 
 
 class reste_cotisations_apres_bic_avant_ba(Variable):
@@ -66,10 +73,11 @@ class reste_cotisations_apres_bic_avant_ba(Variable):
         return max_(
             (
                 individu("cotisations_non_salarie", period)
-                - individu("bic_forfait", period),  # Ne concerne pas les BIC réels
-                0,
-                )
+                - individu("bic_forfait", period)  # Ne concerne pas les BIC réels
+                ),
+            0,
             )
+
 
 
 class reste_cotisations_apres_bic_ba_avant_bnc(Variable):
@@ -83,24 +91,7 @@ class reste_cotisations_apres_bic_ba_avant_bnc(Variable):
         return max_(
             (
                 individu("reste_cotisations_apres_bic_avant_ba", period)
-                - individu("bic_forfait", period),
-                0,
-                )
+                - individu("bic_forfait", period)
+                ),
+            0
             )
-
-# Helpers
-
-def get_multiple_and_plafond_cafat_cotisation(period, parameters):
-    """Renvoie le plafond de la cotisation CAFAT pour l'année revenus donnée."""
-
-    period_plafond = period.start.offset("first-of", "month").offset(11, "month")
-    cafat = parameters(period_plafond).prelevements_obligatoires.prelevements_sociaux.cafat
-    cotisations = parameters(period).prelevements_obligatoires.impot_revenu.revenus_imposables.non_salarie.cotisations
-    if period_plafond.year >= 2023:
-        plafond_cafat = cafat.maladie_retraite.plafond # Donc année revenus 2023
-        multiple = cotisations.plafond_depuis_ir_2024
-    else:
-        plafond_cafat = cafat.autres_regimes.plafond
-        multiple = cotisations.plafond_avant_ir_2024
-
-    return multiple, plafond_cafat
